Created a LocaleAwareTrait, abstracting access to available languages/translations
